libcrux_ml_kem::mlkem1024

Module incremental

Incremental API.

NOTE: This is a non-standard API. Use with caution!

Structs

KeyPairBytes

An encoded, incremental key pair.

KeyPairCompressedBytes

An encoded, compressed, incremental key pair.

PublicKey1

The incremental public key that allows generating [Ciphertext1].

PublicKey2

The incremental public key that allows generating [Ciphertext2].

Constants

COMPRESSED_KEYPAIR_LEN

The size of a compressed key pair in bytes.

Functions

decapsulate_compressed_key

Decapsulate incremental ciphertexts.

decapsulate_incremental_key

Decapsulate incremental ciphertexts.

encaps_state_len

The size of the encaps state in bytes.

encapsulate1

Encapsulate the first part of the ciphertext.

encapsulate2

Encapsulate the second part of the ciphertext.

generate_key_pair

Generate a key pair and write it into key_pair. This uses unpacked keys and does not compress the keys.

generate_key_pair_compressed

Generate a key pair and write it into key_pair. This compresses the keys.

key_pair_compressed_len

The size of the compressed key pair in bytes.

key_pair_len

The size of the key pair in bytes.

pk1

Get the PK1 bytes from the serialized key pair bytes

pk2

Get the PK2 bytes from the serialized key pair bytes

pk1_len

Get the size of the first public key in bytes.

pk2_len

Get the size of the second public key in bytes.

shared_secret_size

The size of the shared secret.

validate_pk

Validate that the two parts pk1 and pk2 are consistent.

validate_pk_bytes

Validate that the two parts pk1 and pk2 are consistent.

Type Aliases

Ciphertext1

Ciphertext 1

Ciphertext2

Ciphertext 2